requirements.txt
cached-property==1.2.0
coloredlogs==1.0.1
docutils==0.12
evdev==0.5.0
executor==4.4
humanfriendly==1.33
lockfile==0.10.2
peewee==2.6.3
proc==0.2.2
pycrypto==2.6.1
PyInstaller==2.1
wheel==0.24.0
zope.interface==4.1.2
